pyproject.toml
[tool.poetry]
name = "jira-cleanup"
version = "0.1.0"
description = "A configurable, policy-based tool for automated Jira ticket governance"
authors = ["Trilliant <info@trilliant.io>"]
readme = "README.md"
packages = [{include = "jiraclean", from = "src"}]
include = [
    "src/jiraclean/prompts/templates/*.yaml",
    "src/jiraclean/prompts/templates/*.yml"
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
jira = ">=3.5.0"
pyyaml = ">=6.0"
python-dotenv = ">=1.0.0"
requests = ">=2.31.0"
typing-extensions = ">=4.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jiraclean = "jiraclean.main:main"
